CPM-137: reorganize front files
