+++ src/Akeneo/UserManagement/Bundle/Resources/public/js/datagrid/action/duplicate-user-action.tsx
@@ -1,7 +1,4 @@
 import React from 'react';
-import {pimTheme} from 'akeneo-design-system';
-import {DependenciesProvider} from '@akeneo-pim-community/legacy-bridge';
-import {ThemeProvider} from 'styled-components';
 import ReactDOM from 'react-dom';
 import {DuplicateUserApp} from '@akeneo-pim-community/user-ui';
 
@@ -16,25 +13,23 @@ class DuplicateUserAction extends AbstractAction {
     const container = document.createElement('div');
     document.body.appendChild(container);
 
-    const closeModal = () => {
+    const closeApp = () => {
       ReactDOM.unmountComponentAtNode(container);
       document.body.removeChild(container);
     };
     const onDuplicateSuccess = (duplicatedUserId: string) => {
-      closeModal();
+      closeApp();
       Routing.redirect(Routing.generate('pim_user_edit', {identifier: duplicatedUserId}));
     };
 
     ReactDOM.render(
-      <DependenciesProvider>
-        <ThemeProvider theme={pimTheme}>
-          <DuplicateUserApp
-            userId={this.model.get(this.propertyName)}
-            onCancel={closeModal}
-            onDuplicateSuccess={onDuplicateSuccess}
-          />
-        </ThemeProvider>
-      </DependenciesProvider>,
+      <>
+        <DuplicateUserApp
+          userId={this.model.get(this.propertyName)}
+          onCancel={closeApp}
+          onDuplicateSuccess={onDuplicateSuccess}
+        />
+      </>,
       container
     );
   }
