+++ src/Akeneo/UserManagement/Bundle/Resources/workspaces/user-ui/src/components/index.ts
@@ -1,0 +1,3 @@
+import {CreateUserForm} from './CreateUserForm';
+
+export {CreateUserForm};
