+++ src/Akeneo/UserManagement/Bundle/Resources/workspaces/user-ui/src/pages/index.ts
@@ -1,0 +1,3 @@
+import DuplicateUser from './DuplicateUser';
+
+export {DuplicateUser};
